requirements-gpu.txt
# GPU deps for RunPod / any CUDA machine.
# Installs onnxruntime-gpu INSTEAD of onnxruntime.
fastapi==0.115.6
uvicorn[standard]==0.32.1
insightface==0.7.3
onnxruntime-gpu==1.20.1
faiss-cpu==1.9.0.post1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1
